Overhaul stuff
New year, new WIMP :]

After working on the alternative trauma kit UI, I've been inspired to
rework WIMP's UI to be more "list-like".
Which means that a lot of things have changed, but WIMP and WOMP are
still here to stay.
This is still a WIP, but I'm hoping to get this finished someday.